requirements_dev.txt
pytest==2.7.2
pytest-django==2.8.0
pytest-cov==1.8.1
